refactor: rename faresByDay
