+++ src/features/authentication/components/index.ts
@@ -1,9 +1,4 @@
 import { CodeFieldComponent, PasswordFieldComponent, TermsFieldComponent, UsernameFieldComponent } from './fields';
 
 // eslint-disable-next-line @typescript-eslint/typedef
-export const COMPONENTS = [
-  CodeFieldComponent,
-  PasswordFieldComponent,
-  TermsFieldComponent,
-  UsernameFieldComponent
-];
+export const COMPONENTS = [CodeFieldComponent, PasswordFieldComponent, TermsFieldComponent, UsernameFieldComponent];
