+++ src/features/authentication/pages/forgot-password/forgot-password.page.html
@@ -24,7 +24,6 @@
       </ng-container>
       <ng-template #defaultText>Envoyer un code de réinitialisation</ng-template>
     </button>
-
   </form>
   <p class="mt-5 text-center">
     <a
