+++ src/features/planning/index.ts
@@ -1,3 +1,3 @@
 export * from './modules/planning.feature.module';
 export * from './providers';
-export * from './reads';
+export * from './queries';
