--- src/features/planning/presentation/fare-by-day.presentation.ts
@@ -1,10 +1,0 @@
-export type FareStatus = 'finished' | 'subcontracted';
-
-export type FareByDayPresentation = {
-  id: string;
-  date: string;
-  distance: string;
-  duration: string;
-  status: FareStatus;
-  startTime: string;
-};
