--- src/features/planning/presentation/index.ts
@@ -1,1 +1,0 @@
-export * from './fare-by-day.presentation';
