+++ src/features/planning/providers/index.ts
@@ -1,1 +1,1 @@
-export * from './read';
+export * from './queries';
