+++ src/features/planning/providers/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './fares-by-day.query.provider';
