--- src/features/planning/providers/read/index.ts
@@ -1,1 +1,0 @@
-export * from './fares-by-day.read.provider';
