+++ src/features/planning/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './fares-by-day/fares-by-day.query';
