--- src/features/planning/reads/index.ts
@@ -1,1 +1,0 @@
-export * from './fares-by-day/fares-by-day.read';
