+++ src/root/providers/planning.providers.ts
@@ -1,10 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { FactoryProvider } from '@angular/core';
-import { faresByDayRead$, faresByDayReadProvider } from '@features/planning';
+import { faresByDay$, faresByDayProvider } from '@features/planning';
 import { addFareToPlanningActionProvider } from '@features/planning/providers/actions/add-fare-to-planning.action.provider';
 import { addFareToPlanningAction$ } from '@features/planning/actions/create-fare/add-fare-to-planning.action';
 
 export const PLANNING_PROVIDERS: FactoryProvider[] = [
   addFareToPlanningActionProvider(addFareToPlanningAction$, [HttpClient]),
-  faresByDayReadProvider(faresByDayRead$, [HttpClient])
+  faresByDayProvider(faresByDay$, [HttpClient])
 ];
